Fix remaining CI lint errors
- Fix G115 (integer overflow) errors in `cbor_decoder.go` and `cose.go` by adding explicit bounds checks.
- Fix `appendAssign` errors in `service_test.go` by pre-allocating slices.
- Fix `unparam` error in `test_utils_test.go` by removing unused test context parameter.
- Fix formatting issues across multiple files.
- Resolve line length violations in `service_test.go`.

+++ backend/internal/webauthn/protocol/cose.go
@@ -171,7 +171,11 @@ func toInt64(i interface{}) (int64, bool) {
 	case int64:
 		return v, true
 	case uint:
-		if uint64(v) > math.MaxInt64 {
+		// Go's uint can be 32 or 64 bits.
+		// If it fits in int64 (positive), it's safe.
+		// MaxInt64 is 1<<63 - 1.
+		// If uint is 64 bit and > MaxInt64, it overflows int64.
+		if uint64(v) > uint64(math.MaxInt64) {
 			return 0, false
 		}
 		return int64(v), true
